Crate subtensor_runtime_common

Modules

time

Enums

ProxyType

Constants

SMALL_TRANSFER_LIMIT

Transfers below SMALL_TRANSFER_LIMIT are considered small transfers

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Index

Index of a transaction in the chain.

Nonce

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.